requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Vector Operations
numpy==1.26.2

# Embedding Service
cohere==4.37

# Temporal Workflow
temporalio==1.5.1

# Async Support
aiofiles==23.2.1

# Retry Logic
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# SDK Dependencies
requests==2.31.0
python-dotenv==1.0.0
